products/codeforge/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
sentence-transformers>=3.4.0
faiss-cpu>=1.9.0
numpy>=1.26.0
pydantic>=2.10.0
gitpython>=3.1.44
torch>=2.5.0
